+++ app/src/androidTest/java/com/example/egobook_frontent/BottomNavigationTest.kt
@@ -7,7 +7,7 @@ import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.example.egobook_frontent.ui.counseling.EgoRoomFragment
-import com.example.egobook_frontent.ui.diary.DiaryFragment
+import com.example.egobook_frontent.ui.diary.view.DiaryFragment
 import com.example.egobook_frontent.ui.home.HomeFragment
 import com.example.egobook_frontent.ui.square.SquareFragment
 import org.assertj.core.api.Assertions.assertThat
