+++ app/src/main/java/com/example/egobook_frontent/ui/diary/adapter/DiaryRVAdapter.kt
@@ -1,9 +1,10 @@
-package com.example.egobook_frontent.ui.diary
+package com.example.egobook_frontent.ui.diary.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.egobook_frontent.databinding.ItemDiaryBinding
+import com.example.egobook_frontent.ui.diary.Diary
 
 class DiaryRVAdapter(private val diaryList: List<Diary>) :
     RecyclerView.Adapter<DiaryRVAdapter.ViewHolder>() {
@@ -12,7 +13,7 @@ class DiaryRVAdapter(private val diaryList: List<Diary>) :
     override fun onCreateViewHolder(
         viewGroup: ViewGroup,
         viewType: Int
-    ): DiaryRVAdapter.ViewHolder {
+    ): ViewHolder {
         val binding: ItemDiaryBinding = ItemDiaryBinding.inflate(
             LayoutInflater.from(viewGroup.context), viewGroup, false
         )
@@ -21,7 +22,7 @@ class DiaryRVAdapter(private val diaryList: List<Diary>) :
 
     }
 
-    override fun onBindViewHolder(holder: DiaryRVAdapter.ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         holder.bind(diaryList[position])
     }
