+++ app/src/main/res/layout/fragment_candler.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.diary.CandlerFragment"
+    tools:context=".ui.diary.view.CandlerFragment"
     android:background="@color/white">
 
     <LinearLayout
