+++ app/src/main/res/layout/fragment_diary_write.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.diary.DiaryWriteFragment"
+    tools:context=".ui.diary.view.DiaryWriteFragment"
     android:background="@color/white">
 
     <!-- 상단 날짜 및 아이콘-->
